Move common cross platform network utils to separate package
