+++ utils/netutil/network_windows.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 The "MysteriumNetwork/node" Authors.
+ * Copyright (C) 2020 The "MysteriumNetwork/node" Authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package userspace
+package netutil
 
 import (
 	"net"
@@ -27,17 +27,12 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-func assignIP(iface string, subnet net.IPNet) error {
+func AssignIP(iface string, subnet net.IPNet) error {
 	out, err := exec.Command("powershell", "-Command", "netsh interface ip set address name=\""+iface+"\" source=static "+subnet.String()).CombinedOutput()
 	return errors.Wrap(err, string(out))
 }
 
-func renameInterface(name, newname string) error {
-	out, err := exec.Command("powershell", "-Command", "netsh interface set interface name=\""+name+"\" newname=\""+newname+"\"").CombinedOutput()
-	return errors.Wrap(err, string(out))
-}
-
-func excludeRoute(ip net.IP) error {
+func ExcludeRoute(ip net.IP) error {
 	gw, err := gateway.DiscoverGateway()
 	if err != nil {
 		return err
@@ -47,7 +42,7 @@ func excludeRoute(ip net.IP) error {
 	return errors.Wrap(err, string(out))
 }
 
-func addDefaultRoute(name string) error {
+func AddDefaultRoute(name string) error {
 	id, gw, err := interfaceInfo(name)
 	if err != nil {
 		return errors.Wrap(err, "failed to get info of interface: "+name)
@@ -59,12 +54,6 @@ func addDefaultRoute(name string) error {
 
 	out, err := exec.Command("powershell", "-Command", "route add 128.0.0.0/1 "+gw+" if "+id).CombinedOutput()
 	return errors.Wrap(err, string(out))
-}
-
-func destroyDevice(name string) error {
-	// Windows implementation is using single device that are reused for the future needs.
-	// Nothing to destroy here.
-	return nil
 }
 
 func interfaceInfo(name string) (id, gw string, err error) {
